Fix unit tests to run with changed Debug format for String
In Rust 1.28, the debug format for the String type changed
to escape more unicode, which broke the event reader tests.
Change to use the String::escape_debug function, to escape
the same set of characters, and also manually add quotation
marks to the format string to avoid changing the expected
results.