Adjust xml lexer to ignore the contents of a <!DOCTYPE> tag
xml-rs does not currently support parsing the contents of the <!DOCTYPE>
tag.  However, the lexer/parser can get confused with an internal
subset being present, as it believes the internal tags are starting
normal tags.

Instead change the lexer to ignore the complete contents inside the
<!DOCTYPE> tag.  This should not affect any external behaviour, as the
parser ignores the contents of the <!DOCTYPE> tag anyways.