Option to allow surrogate pair entity references
Make it possible for the parser to successfully parse an XML document
that contains entity references for surrogate pairs.

A document that contains a character entity reference that resolves to a
surrogate pair – for example `<doc>&#xd83d;&#xdd34;</doc>` – is
*technically* invalid according to the XML specification. However,
replacing these with U+FFFD is both better for the consuming application
(because the application does not need to do some weird pre-processing
of the document before parsing it) and has precedent in that the
standard Go XML parser in its standard configuration replaces these with
U+FFFD.

Add an option – disabled by default – to ParserConfig that allows this
translation when these entity references are encountered.