Force XML document declaration to be output before characters
The 'characters' writer event could not be used to output text after
the document declaration, only before it or after the first XML element.
This change ensures that the document declaration will be output before
the text.  Arbitrary text can still be output at the start of the file,
if needed, by writing directly using the inner_mut method.

Addresses issue #168.