EventWriter: escape '>' in characters
Fixes an issue where the emitter would produce XML which cannot be
parsed back (the reader panics):
- Writer emits the characters `]]>`.
- Reader interprets this as a CDATA suffix instead of the literal `]]>`,
  and panics.

This issue can occur in particular when processing dumps of MediaWiki
markup which uses `[[...]]` to denote links: `Let [[x]]>0`.

A simple fix is to escape all `>` characters in text when emitting XML.
This avoids accidentally forming `]]>` substrings.